requirements.txt
# Core dependencies
qbittorrent-api==2.9.4
pyyaml==6.0
requests==2.28.2

# Optional: for advanced logging
python-json-logger==2.0.7

# Development and testing
pytest==7.3.1
mypy==1.3.0